pull out common code into a util function
